Cleaning up

+++ EulerLib/include/EulerLib.h
@@ -6,7 +6,7 @@
 
 struct SetOfBits;
 
-int fib(int number);
+uint64_t fib(uint64_t number);
 std::string multipleLargeNumbers(std::string numOne, std::string numTwo);
 std::string addLargeNumbers(std::string numberOne, std::string numberTwo);
 std::vector<int> createNumber(std::string number);
@@ -15,22 +15,22 @@ std::vector<uint64_t> getProperDivisors(uint64_t number);
 uint64_t sumOfVector(std::vector<uint64_t> numbers);
 std::vector<std::string> loadNumbers(std::string fileName);
 std::vector<uint64_t> primeFactors(uint64_t number);
-bool isPalindrome(int number);
-bool isDivisibleByRange(int begin, int end, int number);
-int sumOfSquares(int number);
-int squareOfSums(int number);
-bool checkPrime(int n);
-int getPrimeByIndex(int index);
-int collatz(int number);
+bool isPalindrome(uint64_t number);
+bool isDivisibleByRange(uint64_t begin, uint64_t end, uint64_t number);
+uint64_t sumOfSquares(uint64_t number);
+uint64_t squareOfSums(uint64_t number);
+bool checkPrime(uint64_t n);
+uint64_t getPrimeByIndex(uint64_t index);
+uint64_t collatz(uint64_t number);
 std::vector<bool> setBits(std::string bitString);
-std::vector<bool> shiftBits(std::vector<bool> bits, int numberOfBits, int amountToShift);
+std::vector<bool> shiftBits(std::vector<bool> bits, uint64_t numberOfBits, uint64_t amountToShift);
 SetOfBits recursiveShifts(SetOfBits solution);
-std::string makeStringFromNumber(int number);
+std::string makeStringFromNumber(uint64_t number);
 std::vector<std::string> loadTriangle(std::string fileName);
 
 struct SetOfBits
 {
 	std::vector<bool> bits;
-	int numberOfBits = 0;
-	int amountToShift = 0;
+	uint64_t numberOfBits = 0;
+	uint64_t amountToShift = 0;
 };
